Refactor: Seperate Hart Modal
